chore: unify supabase admin client
