+++ lib/supabaseAdmin.js
@@ -1,16 +1,20 @@
 // lib/supabaseAdmin.js
-import { createClient } from "@supabase/supabase-js";
+import { createClient } from '@supabase/supabase-js';
 
-const SUPABASE_URL = process.env.NEXT_PUBLIC_SUPABASE_URL;
-const SUPABASE_SERVICE_ROLE = process.env.SUPABASE_SERVICE_ROLE_KEY;
+// NOTE: Don't throw at build time; Vercel injects envs at runtime for API routes.
+const url = process.env.NEXT_PUBLIC_SUPABASE_URL || '';
+const serviceKey = process.env.SUPABASE_SERVICE_ROLE_KEY || '';
 
-if (!SUPABASE_URL) {
-  console.warn("Missing NEXT_PUBLIC_SUPABASE_URL env var");
-}
-if (!SUPABASE_SERVICE_ROLE) {
-  console.warn("Missing SUPABASE_SERVICE_ROLE_KEY env var");
-}
+let _client = null;
 
-export const supabaseAdmin = createClient(SUPABASE_URL, SUPABASE_SERVICE_ROLE, {
-  auth: { persistSession: false },
-});
+/** Shared Supabase Admin client (service role) for API routes */
+export const supabaseAdmin = (() => {
+  if (!_client) {
+    _client = createClient(url, serviceKey, {
+      auth: { persistSession: false },
+      global: { headers: { 'X-Client-Info': 'listgenie-admin' } },
+    });
+  }
+  return _client;
+})();
+
